server: ensure drain of client connections at the end of decommission
Previously, operators did not have the option to drain late into the decommissioning process,
which caused queries to hang. This was a result of SQL client connections at the end of the
process becoming suddenly unusable. This patch ensures that these client connections on the
decommissioning node are closed before the node is marked as decommissioned.

There is one caveat which is that the draining step of the decommissioning process will be
skipped if the node serving the decommission request is a target. A follow-up PR will fix
this by enforcing that nodes serving requests are not target nodes. Decommission requests
will be routed to other nodes in the cluster that are not targets.

Release note (bug fix): `cockroach node decommission` no longer causes query failure due to
the decommissioning node not closing open SQL connections and still being marked as ready.
The decommissioning process now includes a draining step that fixes this. In other words,
a decommission now automatically drains a node. This also means that running a drain after
a decommission is no longer necessary. It is optional, but recommended, that
`cockroach node drain` is used before `cockroach node decommission` to avoid the
possibility of a disturbance in query performance.