rpc: attempt SRV query on every dial
Previously, SRV resolution (when `experimental-dns-srv` is set)
is only attempted during bootstrapping.
If SRV records don't exist then,
it defaults to 26287 port.
This can cause the node never to be able to join
the cluster.

This PR makes crdb to query for SRV records
on every dial attempt, when `experimental-dns-srv` is set.
If SRV query returns empty, it fallsback to dial
the target directly.

Fixes: CRDB-26962
Release note: None