concurrency: do not partition locks in the lock table by span scope
This patch is entirely a refactor and does not change any
functionality. This is done in preparation for introducing
`LockSpanSets` to track lock spans, which do not make a distinction
between global and local keys (unlike `SpanSets`, which do).

The main changes here are in `lockTableImpl`, which actually stores
locks, and `lockTableGuardImpl` which snapshots the lock table. We
no longer make a distinction between locks on Local and Global keys
when storing them.  The majority of this diff is composed of test file
churn caused because of the printing changes to the lock table.

Informs #102008

Release note: None